scripts/requirements.txt
jsonschema>=4.0.0
pyyaml>=6.0
langdetect>=1.0.9
pandas>=2.0.0
pypdf2>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
tqdm>=4.65.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
